+++ cli/tests/test_bookmark_command.rs
@@ -889,7 +889,6 @@ fn test_bookmark_delete_glob() {
     error: invalid value 'whatever:bookmark' for '<NAMES>...': Invalid string pattern kind `whatever:`
 
     For more information, try '--help'.
-    Hint: Try prefixing with one of `exact:`, `glob:`, `regex:`, or `substring:`
     [EOF]
     [exit status: 2]
     ");
@@ -1129,6 +1128,36 @@ fn test_bookmark_forget_fetched_bookmark() {
     feature1@origin: tyvxnvqr 9175cb32 (empty) another message
     [EOF]
     ");
+
+    // TEST 5: Explicitly naming the remote bookmarks has the same behavior as
+    // --include-remotes (same as test 1)
+    test_env
+        .run_jj_in(
+            &repo_path,
+            ["bookmark", "forget", "feature1", "feature1@origin"],
+        )
+        .success();
+    insta::assert_snapshot!(get_bookmark_output(&test_env, &repo_path), @"");
+    let output = test_env.run_jj_in(&repo_path, ["git", "export"]);
+    insta::assert_snapshot!(output, @"");
+    let output = test_env.run_jj_in(&repo_path, ["git", "import"]);
+    insta::assert_snapshot!(output, @r#"
+    ------- stderr -------
+    Nothing changed.
+    [EOF]
+    "#);
+    insta::assert_snapshot!(get_bookmark_output(&test_env, &repo_path), @"");
+    let output = test_env.run_jj_in(&repo_path, ["git", "fetch", "--remote=origin"]);
+    insta::assert_snapshot!(output, @r#"
+    ------- stderr -------
+    bookmark: feature1@origin [new] tracked
+    [EOF]
+    "#);
+    insta::assert_snapshot!(get_bookmark_output(&test_env, &repo_path), @r#"
+    feature1: tyvxnvqr 9175cb32 (empty) another message
+      @origin: tyvxnvqr 9175cb32 (empty) another message
+    [EOF]
+    "#);
 }
 
 #[test]