+++ CHANGELOG.md
@@ -93,6 +93,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * The new `jj sign` and `jj unsign` commands allow for signing/unsigning commits.
 
+* `jj bookmark forget` can now also accept remote bookmarks as arguments. For
+  instance, `jj bookmark forget bookmark@remote` will forget the target of
+  `bookmark@remote` until the next `jj git fetch` from the remote.
+
 ### Fixed bugs
 
 * `jj status` now shows untracked files under untracked directories.